wpensar/api/requirements.txt
Django==1.8.5
djangorestframework==3.3.2
